Tidy up the webtorrent package, remove buffer
